+++ ExpandWorldSize/ExpandWorldSize.cs
@@ -12,7 +12,7 @@ public class EWS : BaseUnityPlugin
 {
   public const string GUID = "expand_world_size";
   public const string NAME = "Expand World Size";
-  public const string VERSION = "1.10";
+  public const string VERSION = "1.11";
 #nullable disable
   public static ManualLogSource Log;
 #nullable enable
@@ -69,8 +69,7 @@ public class EWS : BaseUnityPlugin
   }
   public void InvokeRegenerate()
   {
-    // Nothing to regenerate because the world hasn't been generated yet.
-    if (WorldGenerator.instance?.m_world?.m_menu != false) return;
+    if (Patcher.IsMenu) return;
     // Debounced for smooth config editing.
     CancelInvoke("Regenerate");
     Invoke("Regenerate", 1.0f);